+++ src/hooks/index.ts
@@ -1,0 +1,4 @@
+import useModal from './useModal';
+import usePopup from './usePopup';
+
+export { useModal, usePopup };
